[REFACTOR] split bivar. and multivar. methods
