+++ mne_connectivity/spectral/epochs_bivariate.py
@@ -1,0 +1,357 @@
+# Authors: Martin Luessi <mluessi@nmr.mgh.harvard.edu>
+#          Denis A. Engemann <denis.engemann@gmail.com>
+#          Adam Li <adam2392@gmail.com>
+#
+# License: BSD (3-clause)
+
+import numpy as np
+
+
+########################################################################
+# Bivariate connectivity estimators
+
+
+class _AbstractConEstBase(object):
+    """ABC for connectivity estimators."""
+
+    def start_epoch(self):
+        raise NotImplementedError('start_epoch method not implemented')
+
+    def accumulate(self, con_idx, csd_xy):
+        raise NotImplementedError('accumulate method not implemented')
+
+    def combine(self, other):
+        raise NotImplementedError('combine method not implemented')
+
+    def compute_con(self, con_idx, n_epochs):
+        raise NotImplementedError('compute_con method not implemented')
+
+
+class _EpochMeanConEstBase(_AbstractConEstBase):
+    """Base class for methods that estimate connectivity as mean epoch-wise."""
+
+    patterns = None
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        self.n_cons = n_cons
+        self.n_freqs = n_freqs
+        self.n_times = n_times
+
+        if n_times == 0:
+            self.csd_shape = (n_cons, n_freqs)
+        else:
+            self.csd_shape = (n_cons, n_freqs, n_times)
+
+        self.con_scores = None
+
+    def start_epoch(self):  # noqa: D401
+        """Called at the start of each epoch."""
+        pass  # for this type of con. method we don't do anything
+
+    def combine(self, other):
+        """Include con. accumated for some epochs in this estimate."""
+        self._acc += other._acc
+
+
+class _CohEstBase(_EpochMeanConEstBase):
+    """Base Estimator for Coherence, Coherency, Imag. Coherence."""
+
+    accumulate_psd = True
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_CohEstBase, self).__init__(n_cons, n_freqs, n_times)
+
+        # allocate space for accumulation of CSD
+        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate CSD for some connections."""
+        self._acc[con_idx] += csd_xy
+
+
+class _CohEst(_CohEstBase):
+    """Coherence Estimator."""
+
+    name = 'Coherence'
+
+    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        csd_mean = self._acc[con_idx] / n_epochs
+        self.con_scores[con_idx] = np.abs(csd_mean) / np.sqrt(psd_xx * psd_yy)
+
+
+class _CohyEst(_CohEstBase):
+    """Coherency Estimator."""
+
+    name = 'Coherency'
+
+    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape,
+                                       dtype=np.complex128)
+        csd_mean = self._acc[con_idx] / n_epochs
+        self.con_scores[con_idx] = csd_mean / np.sqrt(psd_xx * psd_yy)
+
+
+class _ImCohEst(_CohEstBase):
+    """Imaginary Coherence Estimator."""
+
+    name = 'Imaginary Coherence'
+
+    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        csd_mean = self._acc[con_idx] / n_epochs
+        self.con_scores[con_idx] = np.imag(csd_mean) / np.sqrt(psd_xx * psd_yy)
+
+
+class _PLVEst(_EpochMeanConEstBase):
+    """PLV Estimator."""
+
+    name = 'PLV'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_PLVEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # allocate accumulator
+        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        self._acc[con_idx] += csd_xy / np.abs(csd_xy)
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        plv = np.abs(self._acc / n_epochs)
+        self.con_scores[con_idx] = plv
+
+
+class _ciPLVEst(_EpochMeanConEstBase):
+    """corrected imaginary PLV Estimator."""
+
+    name = 'ciPLV'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_ciPLVEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # allocate accumulator
+        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        self._acc[con_idx] += csd_xy / np.abs(csd_xy)
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        imag_plv = np.abs(np.imag(self._acc)) / n_epochs
+        real_plv = np.real(self._acc) / n_epochs
+        real_plv = np.clip(real_plv, -1, 1)  # bounded from -1 to 1
+        mask = (np.abs(real_plv) == 1)  # avoid division by 0
+        real_plv[mask] = 0
+        corrected_imag_plv = imag_plv / np.sqrt(1 - real_plv ** 2)
+        self.con_scores[con_idx] = corrected_imag_plv
+
+
+class _PLIEst(_EpochMeanConEstBase):
+    """PLI Estimator."""
+
+    name = 'PLI'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_PLIEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # allocate accumulator
+        self._acc = np.zeros(self.csd_shape)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        self._acc[con_idx] += np.sign(np.imag(csd_xy))
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        pli_mean = self._acc[con_idx] / n_epochs
+        self.con_scores[con_idx] = np.abs(pli_mean)
+
+
+class _PLIUnbiasedEst(_PLIEst):
+    """Unbiased PLI Square Estimator."""
+
+    name = 'Unbiased PLI Square'
+    accumulate_psd = False
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+        pli_mean = self._acc[con_idx] / n_epochs
+
+        # See Vinck paper Eq. (30)
+        con = (n_epochs * pli_mean ** 2 - 1) / (n_epochs - 1)
+
+        self.con_scores[con_idx] = con
+
+
+class _DPLIEst(_EpochMeanConEstBase):
+    """DPLI Estimator."""
+
+    name = 'DPLI'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_DPLIEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # allocate accumulator
+        self._acc = np.zeros(self.csd_shape)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        self._acc[con_idx] += np.heaviside(np.imag(csd_xy), 0.5)
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+
+        con = self._acc[con_idx] / n_epochs
+
+        self.con_scores[con_idx] = con
+
+
+class _WPLIEst(_EpochMeanConEstBase):
+    """WPLI Estimator."""
+
+    name = 'WPLI'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_WPLIEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # store  both imag(csd) and abs(imag(csd))
+        acc_shape = (2,) + self.csd_shape
+        self._acc = np.zeros(acc_shape)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        im_csd = np.imag(csd_xy)
+        self._acc[0, con_idx] += im_csd
+        self._acc[1, con_idx] += np.abs(im_csd)
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+
+        num = np.abs(self._acc[0, con_idx])
+        denom = self._acc[1, con_idx]
+
+        # handle zeros in denominator
+        z_denom = np.where(denom == 0.)
+        denom[z_denom] = 1.
+
+        con = num / denom
+
+        # where we had zeros in denominator, we set con to zero
+        con[z_denom] = 0.
+
+        self.con_scores[con_idx] = con
+
+
+class _WPLIDebiasedEst(_EpochMeanConEstBase):
+    """Debiased WPLI Square Estimator."""
+
+    name = 'Debiased WPLI Square'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_WPLIDebiasedEst, self).__init__(n_cons, n_freqs, n_times)
+        # store imag(csd), abs(imag(csd)), imag(csd)^2
+        acc_shape = (3,) + self.csd_shape
+        self._acc = np.zeros(acc_shape)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        im_csd = np.imag(csd_xy)
+        self._acc[0, con_idx] += im_csd
+        self._acc[1, con_idx] += np.abs(im_csd)
+        self._acc[2, con_idx] += im_csd ** 2
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+
+        # note: we use the trick from fieldtrip to compute the
+        # the estimate over all pairwise epoch combinations
+        sum_im_csd = self._acc[0, con_idx]
+        sum_abs_im_csd = self._acc[1, con_idx]
+        sum_sq_im_csd = self._acc[2, con_idx]
+
+        denom = sum_abs_im_csd ** 2 - sum_sq_im_csd
+
+        # handle zeros in denominator
+        z_denom = np.where(denom == 0.)
+        denom[z_denom] = 1.
+
+        con = (sum_im_csd ** 2 - sum_sq_im_csd) / denom
+
+        # where we had zeros in denominator, we set con to zero
+        con[z_denom] = 0.
+
+        self.con_scores[con_idx] = con
+
+
+class _PPCEst(_EpochMeanConEstBase):
+    """Pairwise Phase Consistency (PPC) Estimator."""
+
+    name = 'PPC'
+    accumulate_psd = False
+
+    def __init__(self, n_cons, n_freqs, n_times):
+        super(_PPCEst, self).__init__(n_cons, n_freqs, n_times)
+
+        # store csd / abs(csd)
+        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
+
+    def accumulate(self, con_idx, csd_xy):
+        """Accumulate some connections."""
+        denom = np.abs(csd_xy)
+        z_denom = np.where(denom == 0.)
+        denom[z_denom] = 1.
+        this_acc = csd_xy / denom
+        this_acc[z_denom] = 0.  # handle division by zero
+
+        self._acc[con_idx] += this_acc
+
+    def compute_con(self, con_idx, n_epochs):
+        """Compute final con. score for some connections."""
+        if self.con_scores is None:
+            self.con_scores = np.zeros(self.csd_shape)
+
+        # note: we use the trick from fieldtrip to compute the
+        # the estimate over all pairwise epoch combinations
+        con = ((self._acc[con_idx] * np.conj(self._acc[con_idx]) - n_epochs) /
+               (n_epochs * (n_epochs - 1.)))
+
+        self.con_scores[con_idx] = np.real(con)
+
+
+# map names to estimator types
+_CON_METHOD_MAP_BIVARIATE = {
+    'coh': _CohEst, 'cohy': _CohyEst, 'imcoh': _ImCohEst, 'plv': _PLVEst,
+    'ciplv': _ciPLVEst, 'ppc': _PPCEst, 'pli': _PLIEst,
+    'pli2_unbiased': _PLIUnbiasedEst, 'dpli': _DPLIEst, 'wpli': _WPLIEst,
+    'wpli2_debiased': _WPLIDebiasedEst
+}
